Move Policy commands to a separate file
